pdf_to_anomaly_pipeline/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
pdf2image>=1.16.0
poppler-utils
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.25.0
tqdm>=4.62.0
opencv-python>=4.5.0
scipy>=1.7.0
joblib>=1.1.0 